Add sub-items to search notes
Today I noticed that hitting "return" on the Keep action invokes it
directly, which allows for different behavior from hitting "space"
(which switches to suggestion mode). This is fantastic, because it lets
us keep suggestions based on a query while also allowing access to a
menu.

Using the new menu capability, I moved listing all notes by access time
to a sub-item, added a new sub-item to list notes modified in the past
two weeks, and added another sub-item to find and list tags. Pushing
into a particular tag runs a query for all documents containing the tag,
which means it's now possible to browse all notes by tag. Hooray!