Require tags to go on "tags:" line
It turns out that some programming languages (SQL in particular) use `@`
to indicate variables. We _could_ try to parse code blocks out of a file
before looking for tags, but I think it's reasonable to only look for
tags on a specific line, and that's much easier to implement. Now tags
are only recognized when they occur on a line that starts with the
string "tags:". For example:

tags: @idea @whatever