Recents: Order recent notes by atime
Now we use the same listing logic for recent notes and all notes, and
the only difference is an optional cutoff point. This makes it clear
that the only benefit of looking at "recent" notes is that you know that
they've been accessed in the last month. Otherwise, the list is the same
as the top of the list of all notes.